Cron_BigQuery_CloudStorage/requirements.txt
Flask==1.0.2
google-api-python-client==1.7.4
google-cloud-bigquery==1.7.0
